fix: Move EncryptionButton to compose toolbar
