+++ Mail/Views/New Message/ComposeEditor/Mobile/EditorMobileToolbarView.swift
@@ -41,6 +41,7 @@ struct EditorMobileToolbarView: View {
     @ObservedObject var textAttributes: TextAttributes
 
     @Binding var isShowingAI: Bool
+    @Binding var isShowingEncryptStatePanel: Bool
 
     let draft: Draft
     let isEditorFocused: Bool
@@ -54,6 +55,7 @@ struct EditorMobileToolbarView: View {
                     isShowingClassicOptions: $isShowingClassicOptions,
                     isShowingFormattingOptions: $isShowingFormattingOptions,
                     isShowingAI: $isShowingAI,
+                    isShowingEncryptStatePanel: $isShowingEncryptStatePanel,
                     draft: draft,
                     isEditorFocused: isEditorFocused
                 )
@@ -84,6 +86,7 @@ struct EditorMobileToolbarView: View {
     EditorMobileToolbarView(
         textAttributes: TextAttributes(),
         isShowingAI: .constant(false),
+        isShowingEncryptStatePanel: .constant(false),
         draft: Draft(),
         isEditorFocused: true
     )
